+++ test/test_list_databases_response.py
@@ -38,6 +38,7 @@ class TestListDatabasesResponse(unittest.TestCase):
             return ListDatabasesResponse(
                 databases = [
                     hotdata.models.database_summary.DatabaseSummary(
+                        created_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
                         default_catalog = '', 
                         expires_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
                         id = '', 
@@ -48,6 +49,7 @@ class TestListDatabasesResponse(unittest.TestCase):
             return ListDatabasesResponse(
                 databases = [
                     hotdata.models.database_summary.DatabaseSummary(
+                        created_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
                         default_catalog = '', 
                         expires_at = datetime.datetime.strptime('2013-10-20 19:20:30.00', '%Y-%m-%d %H:%M:%S.%f'), 
                         id = '',